api_server/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
redis==5.2.0
httpx==0.27.2
authlib==1.3.2
python-dotenv==1.0.1
itsdangerous==2.2.0
